+++ Lab3/RNGs/Lcg.cs
@@ -5,18 +5,18 @@ namespace Lab3.RNGs
 {
     public class Lcg
     {
-        public int Modulus { get; set; }
-        public int Multiplier { get; set; }
-        public int Increment { get; set; }
-        public int Seed { get; set; }
+        public long Modulus { get; set; }
+        public long Multiplier { get; set; }
+        public long Increment { get; set; }
+        public long Seed { get; set; }
 
-        public int Next()
+        public long Next()
         {
             Seed = (Multiplier * Seed + Increment) % Modulus;
             return Seed;
         }
 
-        public int Next(int last)
+        public long Next(int last)
         {
             Seed = (Multiplier * last + Increment) % Modulus;
             return Seed;
